+++ GameboyArcade/CoreBoy/sound/SoundOutput.cs
@@ -4,6 +4,6 @@ namespace CoreBoy.sound
     {
         void Start();
         void Stop();
-        void Play(int left, int right);
+        void Play(byte left, byte right);
     }
 }
